Greatly simplified gazebo server and gui initialization
The gui when launched needs the server already running. Since launching the gui is optional, the best pattern to create the server is through lazy initialization. In this way, the server is initialized earlier at the first render() call, or alternatively at the first step() when running in headless mode.

When render() is called, the trick that simplifies the logic is running the server as paused in non-blocking mode. The first run() call will unpause the simulation and wait its conclusion.

As a consequence, multiple call to render() are now allowed. It might be useful for opening again the gui from scripting languages like python.