New RobotController ignition plugin
This plugin executes receives dummy action and returns a dummy state. Its purpose is to be associated to a sdf model to expose a gympp::Robot object.

This is the first step to allow creating gym-ignition environments from python. The need of dummy actions / observations will be removed in next commits.