Allow defining the plugin update rate
This information is used to automatically calculate the number of iterations of the simulator.

The user of this class sets the rate of the physics (which might be high) and the rate of the ignition plugins. E.g., if the plugin is a PID controller, it is possible to use a slower rate for the controller while stepping the physics with at a higher frequency.