Improved IgnitionRobot class
- Implemented new joint methods
- Improved existing logic

This class now is similar to the JointController class of gazebo classic. In order to process references, a new update method should be called. It handles automatically the execution rate of the controller.

The ECM components have been updated. They should be more similar to the ones that will be implemented upstream soon.