Removed pimpl from EnvironmentCallbackSingleton and used static attr
This is necessary in release if multiple environments are created in the same process.
The problem is related to the deletion of an environment, that causes the dlclose of the related plugin.
At least from python, the plugin also unloads the shared library of the singleton (shared with other plugins), that causes the destructor to be called, destructing the pimpl object.
In this way, the other plugins that are still running get broken.

Deleting the pimpl and transforming the only attribute to static fixes the situation.
The destructor is still called, but it has no effect, and the state of the class is preserved thanks to the static modifier.